+++ Assets/Scripts/PlayerFall.cs
@@ -3,12 +3,12 @@ using UnityEngine;
 public class PlayerFall : MonoBehaviour {
 
     public PlayerMovement movement;
-    static int highScore = 0;
+    static float highScore = 0f;
 
     void OnCollisionEnter (Collision collisionInfo) {
         if (collisionInfo.collider.tag == "Floor") {
             movement.enabled = false;
-            int score = movement.GetScore();
+            float score = movement.GetScore();
             if (highScore < score)
                 highScore = score;
             FindObjectOfType<GameManager>().EndGame(score, highScore);
